requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
Django==4.2
django-autoslug==1.9.9
django-crispy-forms==2.0
django-jazzmin==2.6.0
idna==2.8
requests==2.21.0
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
urllib3==1.24.3
